% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AskCredentials.R
\name{AskCredentials}
\alias{AskCredentials}
\title{Ask Credentials}
\usage{
AskCredentials()
}
\value{
Vector of 2
}
\description{
Function to get the credentials that will be used to access the Form's API. Generally this function will be called from within GetResponses()
}
\examples{
 AskCredentials()

}
