% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetRespones.R
\name{GetResponses}
\alias{GetResponses}
\title{Get Responses}
\usage{
GetResponses(
  base_url = "https://submit.digital.gov.bc.ca/app/api/v1",
  form_id = NULL,
  api_key = NULL,
  drafts = F,
  deleted = F,
  content.only = T,
  reenter.credentials = F
)
}
\arguments{
\item{base_url}{URL where the form is stored. for the Gov't of BC, the URL is "https://submit.digital.gov.bc.ca/app/api/v1"(default)}

\item{form_id}{if left  (Default), and credentials have not been entered, it will run AskCredentials() to obtain the information. You can find your Form's ID by going to your Form Settings page, and copying it from the URL (after the "=" sign in the URL)}

\item{api_key}{if left  (Default), and credentials have not been entered, it will run AskCredentials() to obtain the information. You can generate an API for the specific form from the "Manage Form" page}

\item{drafts}{Default FALSE, set to TRUE to download forms in draft}

\item{deleted}{Default FALSE, set to TRUE to download deleted forms}

\item{content.only}{Default TRUE. if F, returns a list with the submission data, content type, content disposition, and API status. If FALSE, it will only return the submission data as a data frame.}

\item{reenter.credentials}{Default as FALSE. if TRUE, it will ask you to reenter the Form ID and API KEY, regardless if they already exist in the system}
}
\value{
if content.only= FALSE, returns a list with the submission data, content type, content disposition, and API status. IF False, it will only return the submission data as a data frame.
}
\description{
Get Responses
}
\examples{
#' \dontrun{ GetResponses(form_id = form_id, api_key = api_key, content.only = F)}

}
