% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AuditLog.R
\name{IsAuditLogActive}
\alias{IsAuditLogActive}
\title{Check whether audit logging is active}
\usage{
IsAuditLogActive()
}
\value{
\code{TRUE} if an audit log is active, otherwise \code{FALSE}.
}
\description{
This is most useful in scripts when you want to conditionally add a manual
note using \code{\link[=WriteAuditLog]{WriteAuditLog()}} only when logging is enabled.
}
\examples{
IsAuditLogActive()
}
